Extract task identifier namer into custom class
